requirements.txt
tqdm
pandas
matplotlib
pyyaml
requests
fpdf2
xlsxwriter
pytest
requests-cache
seaborn
black